cloudinary delete api support for more routes

+++ controllers/employeeController.js
@@ -1,6 +1,7 @@
 const catchAsync = require('../utils/catchAsync')
 const AppError = require('../utils/appError')
 const Employee = require('../models/EmployeeModel')
+const cloudinary = require('cloudinary').v2
 
 exports.getAllEmployees = catchAsync(async (req, res, next) => {
     const employees = await Employee.find()
@@ -46,16 +47,30 @@ exports.getLogedInEmployees = catchAsync(async (req, res, next) => {
 exports.updateEmployeeBaseInfo = catchAsync(async (req, res, next) => {
     const { employeeId } = req.params
     const updatedEmployee = await Employee.findOne({ _id: employeeId })
+    const publicId = updatedEmployee.employeeImage.split('/')[7].split('.')[0]
 
-    updatedEmployee.firstName = req.body.firstName || updatedEmployee.firstName
-    updatedEmployee.lastName = req.body.lastName || updatedEmployee.lastName
-    updatedEmployee.birthDate = req.body.birthDate || updatedEmployee.birthDate
-    updatedEmployee.gender = req.body.gender || updatedEmployee.gender
-    updatedEmployee.phone = req.body.phone || updatedEmployee.phone
-    updatedEmployee.role = req.body.role || updatedEmployee.role
-    await updatedEmployee.save({ validateBeforeSave: true })
+    try {
+        updatedEmployee.firstName = req.body.firstName || updatedEmployee.firstName
+        updatedEmployee.lastName = req.body.lastName || updatedEmployee.lastName
+        updatedEmployee.birthDate = req.body.birthDate || updatedEmployee.birthDate
+        updatedEmployee.gender = req.body.gender || updatedEmployee.gender
+        updatedEmployee.phone = req.body.phone || updatedEmployee.phone
+        updatedEmployee.employeeImage = req.files ? req.files.image : updatedEmployee.employeeImage
+        updatedEmployee.role = req.body.role || updatedEmployee.role
 
-    res.status(202).json({
+        await cloudinary.api.delete_resources(publicId, { invalidate: true },
+            function (error, result) {
+                if (error) {
+                    console.log(error)
+                }
+            });
+
+        await updatedEmployee.save({ validateBeforeSave: true })
+    } catch (error) {
+        console.log(error)
+    }
+
+    res.status(200).json({
         message: 'success',
         updatedEmployee
     })
@@ -63,7 +78,25 @@ exports.updateEmployeeBaseInfo = catchAsync(async (req, res, next) => {
 
 exports.deleteEmployee = catchAsync(async (req, res, next) => {
     const { employeeId } = req.params
-    await Employee.findByIdAndDelete(employeeId)
+    const employee = await Employee.findOne({ _id: employeeId })
+    const publicId = employee.employeeImage.split('/')[7].split('.')[0]
+
+    if (!employee) {
+        return next(new AppError('Zaposlenika nije moguće obrisati. Za pomoć kontaktirajte korisničku podršku.', 404))
+    }
+
+    try {
+        await Employee.findByIdAndDelete(employeeId)
+
+        await cloudinary.api.delete_resources(publicId, { invalidate: true },
+            function (error, result) {
+                if (error) {
+                    console.log(error)
+                }
+            });
+    } catch (err) {
+        console.log(err)
+    }
 
     res.status(204).json({
         message: 'success'
